Move fieldmark dialog into CreateNew
